Don't decode the entire Vorbis and FLAC files at load time.
What this change does is makes it so Vorbis and FLAC sound source are
decoder as needed rather than all at once at load time. The rationale
for this change is that it avoids a lot of wasted memory usage.

The downside to this is that it adds latency to audio playback since it
takes longer to mix the audio and send audio data to the device. If
this is unacceptable for an application, they can always do a full
decode themselves and simply create a raw sound source.